requirements.txt
pandas>=1.1.0
numpy>=1.19.0
matplotlib>=3.3.0
scikit-learn>=0.23.0
termcolor>=1.1.0
python-dotenv>=0.14.0
requests>=2.24.0
alpaca-trade-api>=3.0.0
websocket-client>=0.57.0
PyYAML>=5.3.1
tqdm>=4.48.0
joblib>=0.16.0